% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_identify_county_township.R
\name{augment_county_township_using_lon_lat}
\alias{augment_county_township_using_lon_lat}
\title{Identify county and township based on lon lat}
\usage{
augment_county_township_using_lon_lat(df)
}
\arguments{
\item{df}{a data frame with numeric class lon and lat columns}
}
\value{
a data frame augmented with Taiwan's county and township
}
\description{
Identify county and township based on lon lat
}
