% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotExtra.R
\name{axis_x_custom}
\alias{axis_x_custom}
\title{Generate scale_x function with flexible ticks and labels setting}
\usage{
axis_x_custom(scale_x)
}
\arguments{
\item{scale_x}{a scale_x function}
}
\value{

}
\description{
Generate scale_x function with flexible ticks and labels setting
}
\examples{
axis_x_date_custom <- axis_x_custom(scale_x_date)
}
