% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_city_township}
\alias{extract_city_township}
\title{Extract cities and townships from addresses}
\usage{
extract_city_township(address)
}
\arguments{
\item{address}{a character vector of addresses starting from city, then township}
}
\value{
a data frame of city, township, and address
}
\description{
Extract cities and townships from addresses
}
