% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{find_switchingPoint}
\alias{find_switchingPoint}
\title{Get switching points data frame from a data frame with x (data/time), y and y2 (for switching comparison)}
\usage{
find_switchingPoint(.df)
}
\arguments{
\item{.df}{a data frame}
}
\value{
a data frame of switching date-time and intersect y value
}
\description{
Get switching points data frame from a data frame with x (data/time), y and y2 (for switching comparison)
}
\examples{
none
}
