% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinese.R
\name{generateColumn_map_id_en}
\alias{generateColumn_map_id_en}
\title{Generate map_id_en column for a data frame}
\usage{
generateColumn_map_id_en(df, map_id)
}
\arguments{
\item{df}{a data frame}

\item{map_id}{a character of a column in df that represents chinese map_id}
}
\value{
a data frame with extra map_id_en column
}
\description{
Generate map_id_en column for a data frame
}
\examples{
none
}
