% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_overlay}
\alias{geom_overlay}
\title{Enhance geom function with background color (default is white)}
\usage{
geom_overlay(geom, overlayColor = "white")
}
\arguments{
\item{geom}{a ggplot2 geom function}

\item{overlayColor}{default="white"}
}
\value{
an enhanced geom function
}
\description{
Enhance geom function with background color (default is white)
}
\examples{
none
}
