% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotExtra.R
\name{geom_sec_axis}
\alias{geom_sec_axis}
\title{Generate geom function for secondary axis y on the right}
\usage{
geom_sec_axis(geom_fn)
}
\arguments{
\item{geom_fn}{a geom function}
}
\value{

}
\description{
Generate geom function for secondary axis y on the right
}
\examples{
geom_line_right <- geom_sec_axis(geom_line)
}
