% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap.R
\name{geom_sf_overggmap}
\alias{geom_sf_overggmap}
\title{A enhanced geom_sf for overlaying a gg object produced by ggmap2}
\usage{
geom_sf_overggmap(data = data, ...)
}
\arguments{
\item{data}{a sf data object.}

\item{...}{others that pass to regular geom_sf}
}
\value{

}
\description{
A enhanced geom_sf for overlaying a gg object produced by ggmap2
}
\examples{
none.
}
