% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_sf_taiwan}
\alias{geom_sf_taiwan}
\title{Generate Taiwan choropleth map to city or township level without any map supplied}
\usage{
geom_sf_taiwan(
  data,
  map_id,
  ...,
  type = c("縣市", "鄉鎮區"),
  cast2multipolygon = T,
  merge_type = "joint",
  need_background = T,
  background.fill = "#c8c5be",
  background.color = "#c8c5be",
  background.size = 0
)
}
\arguments{
\item{data}{A data frame with map_id to merge with underlying implicit Taiwan map sf object}

\item{map_id}{A character showing the name of the map_id in data input. The data[[map_id]] will be used to join with sf_taiwan_simplied$台灣本島$xxx's map_id column where xxx is determined by type input value.}

\item{...}{other parameters passed to geom_sf}

\item{type}{either "縣市" or "鄉鎮區"}

\item{merge_type}{default="joint". "join" means sf inner_join data, "map" means sf left_join data, "data" means sf right_join data}

\item{need_background}{default is TRUE. If FALSE, just return the choropleth layer.}

\item{background.fill}{the fille of background Taiwan map}

\item{background.color}{the color of background Taiwan map}

\item{background.size}{the size of boundary of background map}
}
\description{
Generate Taiwan choropleth map to city or township level without any map supplied
}
\examples{
ggplot() +
  geom_sf_taiwan(
    data=gg_drug$data,
    map_id="發生地點",
    mapping=aes(fill=案件次數),
    type="鄉鎮區"
  )
}
