% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_switching_regime_dataframe}
\alias{get_switching_regime_dataframe}
\title{Extend a data frame with x(data/time), y/y2 (two numerics) with ymin, ymx and regime}
\usage{
get_switching_regime_dataframe(.df, targetColumns)
}
\arguments{
\item{.df}{a data frame to select x, y and y2 from, where x must have "POSIXct" "POSIXt" class, y and y2 are numeric}

\item{targetColumns}{a named character as c("x"="x in .df", "y"="y in .df", "y2"="y2 in .df") where "zzz in .df" means column name in .df that represents zzz vector for switching regime dataset construction.}
}
\value{

}
\description{
The function augement a dataset with three more columns, ymin, ymax and regime, where ymin and ymax are pmin and pmax of y and y2, regime tells y>=y2 or y<=y2 happens. In addition, interpolated x observations are inserted to prevent void area in geom_ribbon usage when x is discreate
}
\examples{
targetColumns = c("x"="year1",
  "y"="商品及服務輸出佔比", "y2"="商品及服務輸入佔比")
  get_switching_regime_dataframe(.df, targetColumns)
}
