% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photoMeta.R
\name{getMetaDataFromFiles}
\alias{getMetaDataFromFiles}
\title{Extract File metat data using EXIF}
\usage{
getMetaDataFromFiles(files, gpsOnly = TRUE, convertGPS = TRUE)
}
\arguments{
\item{files}{a vector of file paths}

\item{gpsOnly}{default = TRUE}

\item{convertGPS}{default = TRUE, convert GPS from degree-minute-second format to decimal format}
}
\value{
a dataframe, unless gpsOnly = FALSE
}
\description{
Need to install EXIFtools first from https://exiftool.org/
}
\examples{
\dontrun{
  list.files("./photos", full.names = T) |>
  getMetaDataFromFiles(gpsOnly = T) -> dfMeta
}
}
