% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photos.R
\name{getPhotoGPSfromGoogleDriveFolderUrl}
\alias{getPhotoGPSfromGoogleDriveFolderUrl}
\title{Get photo GPS from photo files sitting inside a Google Drive folder}
\usage{
getPhotoGPSfromGoogleDriveFolderUrl(folderUrl, deauth = T)
}
\arguments{
\item{folderUrl}{the url of a google drive folder that has photo files.}

\item{deauth}{default=T, meaning no need for google authentication.}
}
\description{
Get photo GPS from photo files sitting inside a Google Drive folder
}
\examples{
\donntrun{
folderUrl = "https://drive.google.com/drive/u/0/folders/1ldcrE5XGgN96qImQ7oYslTQCG2ccohNd"
folderUrl |>
  getPhotoGPSfromGoogleDriveFolderUrl()
}
}
