% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap.R
\name{ggmap2}
\alias{ggmap2}
\title{A fix to ggmap function that put overly align}
\usage{
ggmap2(map)
}
\arguments{
\item{map}{A map object obtained from ggmap::get_...map()}
}
\value{
a gg object
}
\description{
A fix to ggmap function that put overly align
}
\examples{
none
}
