% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{\%t=\%}
\alias{\%t=\%}
\title{Special function for targets::tar_target(a, fun()) usage}
\usage{
x \%t=\% ...
}
\arguments{
\item{x}{a symbol}

\item{...}{other tar_target arguments}
}
\value{
tar_target call result
}
\description{
Special function for targets::tar_target(a, fun()) usage
}
\examples{
\donotrun{
tar_target(a, fun()) # the same as
a \%t=\% fun()
}
}
