% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hclToHex}
\alias{hclToHex}
\title{Convert h,c,l to hex color code}
\usage{
hclToHex(h, c, l)
}
\arguments{
\item{l}{}
}
\description{
Convert h,c,l to hex color code
}
\examples{
hclToHex(60,40,60) # return #A78D5F
}
