% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{hexToHcl}
\alias{hexToHcl}
\title{Convert hex to h,c,l color code}
\usage{
hexToHcl(hex)
}
\arguments{
\item{hex}{a character vector of hex codes}
}
\value{
a matrix with nrow = hex length
}
\description{
This is a wrapper from https://github.com/allopole/datacolor
}
