% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeup.R
\name{Makeup}
\alias{Makeup}
\title{Collection of ggplot personal designs}
\usage{
Makeup()
}
\value{

}
\description{
Collection of ggplot personal designs
}
\examples{
\donotrun{
makeup = econDV2::Makeup()
plt$theme = makeup$theme()
}
}
