% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{Object}
\alias{Object}
\title{Initiate an object with self saving method.}
\usage{
Object(objectname)
}
\arguments{
\item{objectname}{A symbol or a character}
}
\description{
Initiate an object with self saving method.
}
\examples{
Object(world)
}
