% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Photos.R
\name{Photos}
\alias{Photos}
\title{Create an instance to deal with photo files that has getGPS method to extract GPS locations of photos}
\usage{
Photos(path)
}
\arguments{
\item{path}{a character of either local path or an url linked to a Google drive folder.}
}
\value{
an instance that has getGPS method.
}
\description{
Create an instance to deal with photo files that has getGPS method to extract GPS locations of photos
}
\examples{
\donntrun{
# If supplied a Google Drive folder url:
folderUrl = "https://drive.google.com/drive/u/0/folders/1ldcrE5XGgN96qImQ7oYslTQCG2ccohNd"
folderUrl |> Photos() -> ph
ph$getGPS() -> dfMeta
# or
ph$getGPS(deauth=F) -> dfMeta # if require Google authentication.

# If supplied a local path (must install exiftool first, may only work for MacOS)
folderPath = "./photos"
folderPath |> Photos() -> ph2
ph2$getGPS() -> dfMeta2
}
}
