% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{resize_image}
\alias{resize_image}
\title{Resize image to fit github issue}
\usage{
resize_image(path)
}
\arguments{
\item{path}{a file path to the image}
}
\description{
Resize image to fit github issue
}
\examples{
none
}
