% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_taiwan_simplified}
\alias{sf_taiwan_simplified}
\title{simple feature object of Taiwan.
The data set is obtained from OSM, and
is simplified}
\format{
An object of class \code{list} of length 4.
}
\usage{
sf_taiwan_simplified
}
\description{
simple feature object of Taiwan.
The data set is obtained from OSM, and
is simplified
}
\keyword{datasets}
